requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.7
Cython==3.2.4
idna==3.11
packaging==26.1
pathspec==1.0.4
pillow==12.2.0
python-dotenv==1.2.2
redis==7.4.0
requests==2.33.1
rgbmatrix @ file:///home/dietpi/matrix/rpi-rgb-led-matrix
scikit_build_core==0.12.2
setuptools==82.0.1
spotipy==2.26.0
urllib3==2.6.3
wheel==0.46.3
